// app/application-types.ts
export type ApplicationPayload = {
  carrier: string;
  device: string;
  familyName: string;
  givenName: string;
  familyNameKana: string;
  givenNameKana: string;
  birthDate: string;
  postalCode: string;
  prefecture: string;
  address: string;
  address2: string;
  tel: string;
  email: string;
  website?: string;
};